# Changing toList to toIterator to improve memory optimization and runt…