…ime performance

Signed-off-by: Matt Ahrens <matthewahrens@gmail.com>

Closes #6198 

Tested on set of ~100 log files totally 15GB in size.  Runtime performance numbers:
- With my change: 506.16s user 911.55s system 258% cpu 9:08.09 total
- Without my change: 1707.42s user 1647.09s system 198% cpu 28:07.51 total

Also confirmed that full GC is happening less with change (32 times vs 26 times).